デバイス判別まで
